+++ tests/integration_tests/subroutines.py
@@ -2,8 +2,9 @@
 import matplotlib.pyplot as plt
 
 # local imports (inference engines)
-from probeye.inference.emcee_.solver import run_emcee_solver
-from probeye.inference.torch_.solver import run_pyro_solver
+from probeye.inference.scipy_.solver import ScipySolver
+from probeye.inference.emcee_.solver import EmceeSolver
+from probeye.inference.torch_.solver import PyroSolver
 
 # local imports (post-processing)
 from probeye.postprocessing.sampling import create_pair_plot
@@ -13,7 +14,8 @@ from probeye.postprocessing.sampling import create_trace_plot
 
 def run_inference_engines(problem, true_values=None, n_steps=1000,
                           n_initial_steps=100, n_walkers=20, plot=True,
-                          verbose=True, run_emcee=True, run_torch=True):
+                          verbose=True, run_scipy=True, run_emcee=True,
+                          run_torch=True):
     """
     Runs a requested selection of inference engines on a given problem. This
     function is in a separate file to avoid repeating the same lines of code in
@@ -37,6 +39,9 @@ def run_inference_engines(problem, true_values=None, n_steps=1000,
         If True, the data and the post-processing plots are plotted.
     verbose : bool, optional
         If True, additional information will be printed to the console.
+    run_scipy : bool, optional
+        If True, the problem is solved with scipy (maximum likelihood estimate).
+        Otherwise, no maximum likelihood estimate is derived.
     run_emcee : bool, optional
         If True, the problem is solved with the emcee solver. Otherwise, the
         emcee solver will not be used.
@@ -47,26 +52,42 @@ def run_inference_engines(problem, true_values=None, n_steps=1000,
 
     # this loop avoids to write down the same lines of code for the post-
     # processing for each inference engine again
-    for inference_engine, requested_to_run in {'emcee': run_emcee,
+    for inference_engine, requested_to_run in {'scipy': run_scipy,
+                                               'emcee': run_emcee,
                                                'pyro':  run_torch}.items():
 
-        # solve the problem with emcee if requested
-        if inference_engine == 'emcee':
+        # solve the problem with scipy (max. likelihood) if requested
+        if inference_engine == 'scipy':
+            is_sampling_solver = False
             if requested_to_run:
-                inference_data = run_emcee_solver(
-                    problem, n_walkers=n_walkers, n_steps=n_steps,
-                    n_initial_steps=n_initial_steps, verbose=verbose)
+                scipy_solver = ScipySolver(problem, verbose=verbose)
+                inference_data = scipy_solver.run_max_likelihood(
+                    true_values=true_values)
+            else:
+                # in this case, the engine was not requested to run
+                continue
+
+        # solve the problem with emcee if requested
+        elif inference_engine == 'emcee':
+            is_sampling_solver = True
+            if requested_to_run:
+                emcee_solver = EmceeSolver(problem, verbose=verbose)
+                inference_data = emcee_solver.run_mcmc(
+                    n_walkers=n_walkers, n_steps=n_steps,
+                    n_initial_steps=n_initial_steps)
             else:
                 # in this case, the engine was not requested to run
                 continue
 
         # solver the problem with pyro/torch if requested
         elif inference_engine == 'pyro':
+            is_sampling_solver = True
             if requested_to_run:
                 n_walkers_used = 1  # getting errors when trying to use more
-                inference_data = run_pyro_solver(
-                    problem, n_walkers=n_walkers_used, n_steps=n_steps,
-                    n_initial_steps=n_initial_steps, verbose=verbose)
+                pyro_solver = PyroSolver(problem, verbose=verbose)
+                inference_data = pyro_solver.run_mcmc(
+                    n_walkers=n_walkers_used, n_steps=n_steps,
+                    n_initial_steps=n_initial_steps)
             else:
                 # in this case, the engine was not requested to run
                 continue
@@ -81,9 +102,10 @@ def run_inference_engines(problem, true_values=None, n_steps=1000,
         # each inference engine; also note, that the plots are not so much
         # intended for automatic testing, as for manually running the script
         if plot:
-            create_pair_plot(inference_data, problem,
-                             true_values=true_values, show=False)
-            create_posterior_plot(inference_data, problem,
-                                  true_values=true_values, show=False)
-            create_trace_plot(inference_data, problem, show=False)
-            plt.show()  # shows all plots at once due to 'show=False' above
+            if is_sampling_solver:
+                create_pair_plot(inference_data, problem,
+                                 true_values=true_values, show=False)
+                create_posterior_plot(inference_data, problem,
+                                      true_values=true_values, show=False)
+                create_trace_plot(inference_data, problem, show=False)
+                plt.show()  # shows all plots at once due to 'show=False' above