+++ tests/integration_tests/test_spatial_correlation.py
@@ -27,7 +27,7 @@ from probeye.definition.noise_model import NormalNoiseModel
 
 # local imports (testing related)
 from tests.integration_tests.subroutines import run_inference_engines
-from probeye.inference.emcee_.correlation_models import \
+from probeye.inference.scipy_.correlation_models import \
     SpatialExponentialCorrelationModel
 
 
@@ -35,7 +35,8 @@ class TestProblem(unittest.TestCase):
 
     def test_spatial_correlation(self, n_steps=200, n_initial_steps=100,
                                  n_walkers=20, plot=False, verbose=False,
-                                 run_emcee=True, run_torch=False):
+                                 run_scipy=True, run_emcee=True,
+                                 run_torch=False):
         """
         Integration test for the problem described at the top of this file.
 
@@ -54,6 +55,9 @@ class TestProblem(unittest.TestCase):
             generated plots are closed.
         verbose : bool, optional
             If True, additional information will be printed to the console.
+        run_scipy : bool, optional
+            If True, the problem is solved with scipy (maximum likelihood est).
+            Otherwise, no maximum likelihood estimate is derived.
         run_emcee : bool, optional
             If True, the problem is solved with the emcee solver. Otherwise,
             the emcee solver will not be used.
@@ -61,6 +65,10 @@ class TestProblem(unittest.TestCase):
             If True, the problem is solved with the pyro/torch_ solver.
             Otherwise, the pyro/torch_ solver will not be used.
         """
+
+        if run_torch:
+            raise RuntimeError("The pyro-solver is not available yet for "
+                               "forward models including correlations.")
 
         # ==================================================================== #
         #                          Set numeric values                          #
@@ -213,7 +221,8 @@ class TestProblem(unittest.TestCase):
         run_inference_engines(problem, true_values=true_values, n_steps=n_steps,
                               n_initial_steps=n_initial_steps,
                               n_walkers=n_walkers, plot=plot, verbose=verbose,
-                              run_emcee=run_emcee, run_torch=run_torch)
+                              run_scipy=run_scipy, run_emcee=run_emcee,
+                              run_torch=run_torch)
 
 if __name__ == "__main__":
     unittest.main()